requirements.txt
matplotlib==3.7.3
numpy==1.24.3
open3d==0.17.0
Pillow==10.0.1
# plotly # should install plotly-orca from conda
plyfile==1.1
pyhocon==0.3.57
PyMaxflow==1.3.0
pytorch3d==0.7.5
scipy==1.15.2
skimage==0.18.0
tensorboardX==2.6.2.2
# torch==2.0.1 # Install with instructions from https://pytorch.org/get-started/locally/
tqdm==4.65.0
trimesh==4.0.0
